+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
+### Changed
+- Made Sessions Cookies Private/Encypted for confidentiality, integrity, and authenticity.
 
 ## 2.0.1 (18. May, 2022)
 ### Fixed